Make text editor immutable? (description)
Not sure how this code is working. I wrote it in one sitting on a whim just to try out immutability with the text editor.

I need to sleep and then revisit this code. A lot of the code is duplicated throughout.

I'm commiting this because it seems like an incredibly important step in the right direction.

I kept getting 'enumeration was modified' exceptions while using the IDE itself.